Implemented JPA-based AccountRepository, with Account now being an Entity.
Accounts REST Controller test uses a "slice" (only @WebMvcTest with AccountsController.class) to test only the Controller, using a mock repository, instead of the fake in-memory repository used before. Since the InMemory repository does not fully implement the new AccountRepository (from JpaRepository), and it's too large of an interface to fake, we'll mock out the repository and force it to return a known Account object.

Accounts REST Integration test uses the full @SpringBootTest annotation that has Spring Boot scan everything, and @AutoConfigureMockMvc to prepare the MockMvc for use. Note that this is still mocking the MVC, so the embedded container does not actually start. Note that because the default configuration, along with the Data Loader, get picked up and run, we're actually testing against that data.

See https://springframework.guru/unit-testing-spring-mvc-spring-boot-1-4-part-1/ for some examples of this, especially the Summary which differentiates between @WebMvcTest and @SpringBootTest. Also see